@@ -7,9 +7,12 @@ license = "MIT"
 package-mode = false
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 scrapy = "^2.11.0"
 typing-extensions = "^4.9.0"
+fastapi = "^0.109.0"
+uvicorn = "^0.27.0"
+pydantic = "^2.6.0"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.8.0"
@@ -19,19 +22,32 @@ pytest-asyncio = "^0.23.5"
 pytest-cov = "^4.1.0"
 black = "^24.2.0"
 pre-commit = "^3.6.0"
+types-requests = "^2.32.0.20241016"
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 check_untyped_defs = true
 mypy_path = "iching/stubs"
+ignore_missing_imports = true
+# plugins = ["pydantic.mypy"]
+follow_imports = "silent"
+implicit_reexport = true
+namespace_packages = true
+explicit_package_bases = true
+
+# [tool.pydantic-mypy]
+# init_forbid_extra = true
+# init_typed = true
+# warn_required_dynamic_aliases = true
+# warn_untyped_fields = true
 
 [tool.ruff]
 select = [
