@@ -1,5 +1,5 @@
 import random
-from typing import List, Optional, Dict, Union, TypedDict, Mapping
+from typing import List, Optional, TypedDict
 
 class HexagramReading(TypedDict):
     name: str
@@ -7,7 +7,7 @@ class HexagramReading(TypedDict):
     description: str
     judgment: str
     image: str
-    lines: Dict[str, str]
+    lines: dict[str, str]
 
 class Hexagram:
     def __init__(self) -> None:
@@ -58,24 +58,4 @@ class Hexagram:
         relating_hexagram.lines = relating_lines
         relating_hexagram.number = relating_hexagram._calculate_hexagram_number()
 
-        return relating_hexagram
-
-class ReadingInterpreter:
-    def __init__(self, readings: Dict[int, HexagramReading]):
-        self.readings = readings
-
-    def get_reading(self, hexagram_number: int, changing_lines: Optional[List[int]] = None) -> HexagramReading:
-        reading = self.readings.get(hexagram_number)
-        if not reading:
-            raise KeyError(f"Reading not found for hexagram number {hexagram_number}")
-
-        if changing_lines:
-            lines_dict = reading.get("lines", {})
-            transformed_lines: Dict[str, str] = {}
-            for line_number in changing_lines:
-                line_meaning = lines_dict.get(str(line_number))
-                if line_meaning:
-                    transformed_lines[str(line_number)] = line_meaning
-            reading["lines"] = transformed_lines
-
-        return reading
+        return relating_hexagram 
